use composer to manage dependencies
It's not common practice to also commit the vendor folder to git. But
this way it's easier for people to just checkout the project and run it.

The composer.json file is very verbose in this case because none of the
dependencies are composer projects. Thus the github urls needed to be
added explicitly here.

The next step would be to package the application as a phar file for end
users and to remove the vendor folder from git.

For building phars see:
http://box-project.github.io/box2
https://moquet.net/blog/distributing-php-cli/
http://yet.another.linux-nerd.com/blog/create-a-simple-phar-command-line-project-based-on-symfony-components